+++ core/shared/src/main/scala/zio/prelude/AssociativeCompose.scala
@@ -17,50 +17,14 @@ trait AssociativeCompose[:=>[-_, +_]] {
 
 object AssociativeCompose {
 
-  implicit val FunctionBothEitherIdentityCompose
-    : ApplicationCompose[Function, ({ type lambda[+f, +s] = (f, s) })#lambda, Function]
-      with BothCompose[Function, ({ type lambda[+f, +s] = (f, s) })#lambda]
-      with EitherCompose[Function, ({ type lambda[+l, +r] = Either[l, r] })#lambda]
-      with IdentityCompose[Function] =
-    new ApplicationCompose[Function, ({ type lambda[+f, +s] = (f, s) })#lambda, Function]
-      with BothCompose[Function, ({ type lambda[+f, +s] = (f, s) })#lambda]
-      with EitherCompose[Function, ({ type lambda[+l, +r] = Either[l, r] })#lambda]
-      with IdentityCompose[Function] {
+  implicit val FunctionIdentityCompose: IdentityCompose[Function] = new IdentityCompose[Function] {
 
-      def identity[A]: A => A = scala.Predef.identity
+    def identity[A]: A => A = scala.Predef.identity
 
-      def compose[A, B, C](bc: B => C, ab: A => B): A => C =
-        bc.compose(ab)
+    def compose[A, B, C](bc: B => C, ab: A => B): A => C =
+      bc.compose(ab)
 
-      override def fromFirst[A]: Function[(A, Any), A] = _._1
-
-      override def fromSecond[B]: Function[(Any, B), B] = _._2
-
-      override def toBoth[A, B, C](a2b: Function[A, B])(a2c: Function[A, C]): Function[A, (B, C)] = { a =>
-        (a2b(a), a2c(a))
-      }
-
-      override def application[A, B]: Function[(Function[A, B], A), B] = { case (a2b, a) =>
-        a2b(a)
-      }
-
-      override def curry[A, B, C](f: Function[(A, B), C]): Function[A, Function[B, C]] = { a => b =>
-        f((a, b))
-      }
-
-      override def uncurry[A, B, C](g: Function[A, Function[B, C]]): Function[(A, B), C] = { case (a, b) =>
-        g(a)(b)
-      }
-
-      override def toLeft[A]: Function[A, Either[A, Nothing]] = Left(_)
-
-      override def toRight[B]: Function[B, Either[Nothing, B]] = Right(_)
-
-      override def fromEither[A, B, C](a2c: => Function[A, C])(b2c: => Function[B, C]): Function[Either[A, B], C] = {
-        case Left(a)  => a2c(a)
-        case Right(b) => b2c(b)
-      }
-    }
+  }
 }
 
 trait AssociativeComposeSyntax {
