+++ core/shared/src/main/scala/zio/prelude/package.scala
@@ -4,21 +4,18 @@ import com.github.ghik.silencer.silent
 import zio.test.{ TestResult, assert }
 
 package object prelude
-    extends ApplicationComposeSyntax
-    with Assertions
+    extends Assertions
     with AssociativeSyntax
     with AssociativeBothSyntax
     with AssociativeComposeSyntax
     with AssociativeEitherSyntax
     with AssociativeFlattenSyntax
-    with BothComposeSyntax
     with CommutativeBothSyntax
     with CommutativeEitherSyntax
     with CovariantSyntax
     with ContravariantSyntax
     with DebugSyntax
     with DivariantSyntax
-    with EitherComposeSyntax
     with EqualSyntax
     with HashSyntax
     with IdExports
@@ -106,13 +103,9 @@ package object prelude
     type Applicative[F[+_]]         = Covariant[F] with IdentityBoth[F]
     type InvariantApplicative[F[_]] = Invariant[F] with IdentityBoth[F]
 
-    type Category[:=>[-_, +_]]                                           = IdentityCompose[:=>]
-    type CartesianCategory[:=>[-_, +_], :*:[+_, +_]]                     = Category[:=>] with BothCompose[:=>, :*:]
-    type ClosedCartesianCategory[:=>[-_, +_], :*:[+_, +_], :-->[-_, +_]] =
-      CartesianCategory[:=>, :*:] with ApplicationCompose[:=>, :*:, :-->]
-    type CoCartesianCategory[:=>[-_, +_], :+:[+_, +_]]                   = Category[:=>] with EitherCompose[:=>, :+:]
-    type Profunctor[:=>[-_, +_]]                                         = Divariant[:=>]
-    type Bifunctor[:=>[+_, +_]]                                          = Bicovariant[:=>]
+    type Category[:=>[-_, +_]]   = IdentityCompose[:=>]
+    type Profunctor[:=>[-_, +_]] = Divariant[:=>]
+    type Bifunctor[:=>[+_, +_]]  = Bicovariant[:=>]
   }
 
   /**
