+++ experimental/shared/src/main/scala/zio/prelude/experimental/ApplicationCompose.scala
@@ -1,4 +1,5 @@
 package zio.prelude
+package experimental
 
 trait ApplicationCompose[:=>[-_, +_], :*:[+_, +_], :-->[-_, +_]] extends BothCompose[:=>, :*:] {
   def application[A, B]: ((A :--> B) :*: A) :=> B
