+++ experimental/shared/src/main/scala/zio/prelude/experimental/BothCompose.scala
@@ -1,4 +1,5 @@
 package zio.prelude
+package experimental
 
 trait BothCompose[:=>[-_, +_], :*:[+_, +_]] extends AssociativeCompose[:=>] {
   def fromFirst[A]: (A :*: Any) :=> A
@@ -16,6 +17,38 @@ trait BothCompose[:=>[-_, +_], :*:[+_, +_]] extends AssociativeCompose[:=>] {
 
     law1 && law2 && law3
   }
+}
+
+object BothCompose {
+
+  implicit val FunctionBothEitherIdentityCompose
+    : ApplicationCompose[Function, ({ type lambda[+f, +s] = (f, s) })#lambda, Function] =
+    new ApplicationCompose[Function, ({ type lambda[+f, +s] = (f, s) })#lambda, Function] {
+
+      override def compose[A, B, C](bc: B => C, ab: A => B): A => C =
+        bc.compose(ab)
+
+      override def fromFirst[A]: Function[(A, Any), A] = _._1
+
+      override def fromSecond[B]: Function[(Any, B), B] = _._2
+
+      override def toBoth[A, B, C](a2b: Function[A, B])(a2c: Function[A, C]): Function[A, (B, C)] = { a =>
+        (a2b(a), a2c(a))
+      }
+
+      override def application[A, B]: Function[(Function[A, B], A), B] = { case (a2b, a) =>
+        a2b(a)
+      }
+
+      override def curry[A, B, C](f: Function[(A, B), C]): Function[A, Function[B, C]] = { a => b =>
+        f((a, b))
+      }
+
+      override def uncurry[A, B, C](g: Function[A, Function[B, C]]): Function[(A, B), C] = { case (a, b) =>
+        g(a)(b)
+      }
+
+    }
 }
 
 trait BothComposeSyntax {
